Implement Onbording page
Base on the details provided for this exercice
and base on the provided mock from Meta,
This page has been implemented accordingly

It could be great to revisit this and implement
unittest if possible. It is imprtant to check
the colors, the text font and style match the
expected result.